+++ CHANGELOG.md
@@ -12,15 +12,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `Icon` component added to the core package [#103](https://github.com/raster-foundry/blasterjs/pull/103)
 - `Box` component added to the core package [#112](https://github.com/raster-foundry/blasterjs/pull/112)
 - `Text` component added to the core package [#114](https://github.com/raster-foundry/blasterjs/pull/114)
+- 'GlobalStyle' component added to the core package [#115](https://github.com/raster-foundry/blasterjs/pull/115)
 
 ### Changed
 
 - Changed theme file `radius` to `radii` (breaking change for custom themes)
 - Normalized usage of styled-system across every component
+- Renamed `Link` to `A`.
 
 ### Deprecated
 
 ### Removed
+
+- Removed reset.css.
 
 ### Fixed